Removed all groups of three
Noticed last night on the server anytime a unit was stuck on a bridge, a house, they would separate from the group, and there would be stutters, and ground unit warping. I decreased all units to max of two, if this helps we may want to consider not spawning groups.